% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_df}
\alias{count_df}
\title{Gene expression profile of 293T cell line}
\format{
A data frame with 27165 rows and 4 variables:
\describe{
  \item{Ctrl_rep1}{gene expression of control WT 293T cell, replication 1}
  \item{Ctrl_rep2}{gene expression of control WT 293T cell, replication 2}
  \item{METTL3_KD_rep1}{gene expression of control WT 293T cell, replication 1}
  \item{METTL3_KD_rep2}{gene expression of control WT 293T cell, replication 2}
}
}
\usage{
count_df
}
\description{
Two replications for WT and METTL3 KD 293T cell line, the counts in the dataframe are results of FeatureCounts.
}
\keyword{datasets}
