% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_Deseq2.R
\name{deg_deseq2}
\alias{deg_deseq2}
\title{Differential expression analysis using DESeq2}
\usage{
deg_deseq2(
  dt,
  control_label,
  control_counts,
  treatment_lable,
  treatment_counts
)
}
\arguments{
\item{dt}{A data frame, rows are gene expression (counts),columns are sample names; the order of columns must be
control samples followed by treatment samples.}

\item{control_label}{A character vector, name for control samples,such as "ctrl".}

\item{control_counts}{A numeric vector, how many samples are control.}

\item{treatment_lable}{A character vector, name for treatment samples,such as "KD"(Knockdown).}

\item{treatment_counts}{A numeric vector, how many samples are treatment.}
}
\value{
Return a result dataframe from a DESeq analysis giving base means across samples, log2 fold changes,
    standard errors, test statistics, p-values and adjusted p-values
}
\description{
Differential expression analysis using DESeq2
}
\examples{
res <- deg_deseq2(count_df,control_label="ctrl",control_counts=2,treatment_lable="KD",treatment_counts=2)
}
