% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_gsea.R
\name{do_gsea}
\alias{do_gsea}
\title{Do gsea}
\usage{
do_gsea(dt, gene_set, gene_id, stat)
}
\arguments{
\item{dt}{dataframe contain gene id (gene_id) and rank statistics (stat)}

\item{gene_set}{pathway information, the path of gmt file}

\item{gene_id}{colname of gene_id}

\item{stat}{colname of stat}
}
\value{
ES
}
\description{
Do gsea
}
\examples{
res <- deg_deseq2(count_df,control_label="ctrl",control_counts=2,treatment_lable="KD",treatment_counts=2,parallel = TRUE,ncores = 4)
res$gene <- rownames(res)
gmt.file <- system.file("extdata", "h.all.v7.4.symbols.gmt", package="EasyBioinfo")
gsea_res <- do_gsea(res,gene_set = gmt.file,gene_id = "gene",stat = "stat")
}
