% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_gsva.R
\name{do_gsva}
\alias{do_gsva}
\title{Do gsva}
\usage{
do_gsva(exp_dt, gmt_path, exp_type = "array", ...)
}
\arguments{
\item{exp_dt}{A normalized gene expression dataset, genes corresponding to rows and samples corresponding to columns.}

\item{gmt_path}{Pathway gmt file path}

\item{exp_type}{expression type, can be "array" or "RNA-seq"}

\item{...}{other param for gsva}
}
\value{
gene-set-by-sample expression data matrix
}
\description{
Do gsva
}
\examples{
res <- do_gsva(as.matrix(tpm_gene_log2),"inst/extdata/h.all.v7.4.symbols.gmt",exp_type = "array",parallel.sz=4)
}
