% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcode.R
\name{getcode}
\alias{getcode}
\title{get cancer type code}
\usage{
getcode(
  x,
  dt = "cancer_type_code",
  column_need = "cancer_type",
  column_select = "cancer_code"
)
}
\arguments{
\item{x}{}
}
\value{
code
}
\description{
get cancer type code
}
\examples{
getcode("01")
}
