% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpep.R
\name{rpep}
\alias{rpep}
\title{Get random peptides from protein sequences}
\usage{
rpep(dt, mer_len, random_num)
}
\arguments{
\item{dt}{A dataframe with at least 1 column containg protein sequence}

\item{mer_len}{An integer number, the length of random peptide}

\item{random_num}{An integer number, the number of random peptides needed}
}
\value{
A dataframe with a column named random_pep, containg the random peptides of sepecific length extracted from given proteins.
}
\description{
Get random peptides from protein sequences
}
\examples{
res <- rpep(uniprot_protein_test,mer_len=9,random_num=100)


}
