% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_survival}
\alias{sample_survival}
\title{Survival data of 1000 samples in TCGA project}
\format{
A data frame with 1000 rows and 6 variables:
\describe{
  \item{sample}{TCGA sample barcode}
  \item{age}{age of samples}
  \item{gender}{gender of samples; Male/Female}
  \item{Tumor_stage}{Tumor stages}
  \item{OS}{OS status}
  \item{OS.time}{Survial time}
}
}
\usage{
sample_survival
}
\description{
Survival data of 1000 samples in TCGA project
}
\keyword{datasets}
