% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survial_analysis.R
\name{show_km}
\alias{show_km}
\title{Wrapper function of plot KM curve by survial and survminer packages}
\usage{
show_km(data, var, ...)
}
\arguments{
\item{data}{A dataframe, containing survial status (as "OS" 0=alive, 1=dead) and survial time (as "OS.time")}

\item{var}{A character vector or column name (not need been quoted), varible to separate samples}

\item{...}{other arguments passed to survminer::ggsurvplot}
}
\value{
ggsurvplot object
}
\description{
Wrapper function of plot KM curve by survial and survminer packages
}
\examples{
show_km(sample_survival,"gender")
## or
show_km(sample_survival,gender)
}
