% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniprot_protein_test}
\alias{uniprot_protein_test}
\title{Protein sequence from Uniprot database (only take 50 proteins for example)}
\format{
A data frame with 50 rows and 3 variables:
\describe{
  \item{id}{Uniport protein accession ID}
  \item{len}{Length of the sequence}
  \item{seq}{Sequence of the protein}
}
}
\source{
\url{https://www.uniprot.org/downloads}
}
\usage{
uniprot_protein_test
}
\description{
A dataframe containing sequences of 50 proteins
}
\keyword{datasets}
